auto battle persistence support and perf improvement
