+++ .gitignore
@@ -8,6 +8,7 @@ __pycache__/
 
 # Distribution / packaging
 .Python
+.history/
 build/
 develop-eggs/
 dist/
