+++ README.md
@@ -19,7 +19,8 @@
 
 **3：强烈推荐使用远程角色开怪**
 
-**4：战斗为游戏自带的自动战斗，确保你的队伍有足够实力平推小怪**
+**4：战斗为游戏自带的自动战斗，确保你的队伍有足够实力平推小怪** 
+(如启用了沿用自动战斗设定，请把config.json里的 "auto_battle_persistence" 改成 1) 
 
 **5：如果脚本无反应，请使用管理员权限运行**
