+++ config.json
@@ -1,0 +1,4 @@
+{
+    "real_width": 1295,
+    "auto_battle_persistence": 0
+}
