--- real_width.json
@@ -1,3 +1,0 @@
-{
-    "real_width": 1295
-}
